getting into the weeds of compilation of larger projects....
