+++ src/evaluation.cpp
@@ -1,0 +1,9 @@
+#include "evaluation.h"
+
+int count_material(const Board* bo){
+    auto turn = bo->get_turn();
+    
+    return turn;
+}
+
+
